+++ app/src/main/java/com/wepli/app/navigation/NavHost.kt
@@ -1,8 +1,5 @@
 package com.wepli.app.navigation
 
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -29,8 +26,8 @@ import com.wepli.feature.photocard.result.navigation.photoCardResultGraph
 import com.wepli.feature.photocard.result.navigation.navigateToPhotoCardResult
 import com.wepli.feature.song.info.navigation.navigateToSongInfo
 import com.wepli.feature.song.info.navigation.songInfoGraph
-import com.wepli.mypage.devmode.network.navigation.navigateToNetworkLogMain
-import com.wepli.mypage.devmode.network.navigation.networkLogMainGraph
+import com.wepli.mypage.devmode.network.main.navigation.navigateToNetworkLogMain
+import com.wepli.mypage.devmode.network.main.navigation.networkLogMainGraph
 import com.wepli.playlist.navigation.navigateToPlaylistDetail
 import com.wepli.playlist.navigation.playlistDetailGraph
 import com.wepli.relaylist.navigation.navigateToRelaylistDetail
