+++ feature/mypage/src/debug/kotlin/com/wepli/mypage/devmode/network/main/navigation/NetworkLogNavigation.kt
@@ -1,9 +1,9 @@
-package com.wepli.mypage.devmode.network.navigation
+package com.wepli.mypage.devmode.network.main.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.wepli.mypage.devmode.network.screen.NetworkLogDebugScreenRoute
+import com.wepli.mypage.devmode.network.main.screen.NetworkLogDebugScreenRoute
 import com.wepli.navigator.feature.mypage.NetworkLogRoute
 
 fun NavController.navigateToNetworkLogMain() {
